Allow suppression of generic response type for multiple content types
When an operation returns multiple content types the return type is
removed which result in a generic java.lang.Object return type. While
this might be useful for operations which return completely different
responses it isn't useful for cases where there is effectively one
response structure.

By allowing the suppression of the generic response type for multiple
content types both the current behaviour is supported by default and the
alternative behaviour can be enabled with a property.