Fix enum classes in native image, annotate as Introspected
Model classes are annotated as Introspected in order to generate the
appropriate metadata for Micronaut and the GraalVM native image
compiler. This isn't the case for enum classes, which isn't a problem
for Micronaut running in a standard JVM but results in errors when used
in a GraalVM native image.

By annotating enum classes as Introspected they can be used in both a
standard JVM and in a GraalVM native image.