Support adding items to a list or map. Fixes #64
List and maps are initialized with null which makes it cumbersome to add
items programmatically because of the mandatory null checks.

By adding a specific method for adding an item to a list or a map,
similar to how is done for the native Java generator, the null checks
are no longer required.